+++ Coimbra.Editor/Utilities/FrameworkEditorUtility.cs
@@ -167,7 +167,7 @@ namespace Coimbra.Editor
 
         private static void OnPostprocessAllAssets(string[] importedAssets, string[] deletedAssets, string[] movedAssets, string[] movedFromAssetPaths)
         {
-            using (SharedPool.Pop(out List<UnityEngine.Object> pooledList))
+            using (SharedManagedPool.Pop(out List<UnityEngine.Object> pooledList))
             {
                 pooledList.Clear();
                 pooledList.AddRange(PlayerSettings.GetPreloadedAssets());