+++ Coimbra/ScriptableSettings.cs
@@ -215,7 +215,7 @@ namespace Coimbra
             }
             else
             {
-                using (SharedPool.Pop(out List<Object> pooledList))
+                using (SharedManagedPool.Pop(out List<Object> pooledList))
                 {
                     pooledList.Clear();
                     pooledList.AddRange(UnityEditor.PlayerSettings.GetPreloadedAssets());
@@ -263,7 +263,7 @@ namespace Coimbra
 
         private void EnsurePreload(bool withWarning)
         {
-            using (SharedPool.Pop(out List<Object> pooledList))
+            using (SharedManagedPool.Pop(out List<Object> pooledList))
             {
                 pooledList.Clear();
                 pooledList.AddRange(UnityEditor.PlayerSettings.GetPreloadedAssets());